+++ inst/NEWS.md
@@ -1,10 +1,16 @@
 __MEDseq: Mixtures of Exponential-Distance Models with Covariates__   
 ===================================================================
 
+### New Features
+* `MEDseq_stderr` is provided for computing the standard errors of the  
+  coefficients for the covariates in the gating network via either the  
+  weighted likelihood bootstrap or jackknife methods.
+* Small robustifications in the presence of empty components.
+* Fixed bug related to the colours used in the vignette plots.
+
 ## MEDseq v1.0.1 - (_2<sup>nd</sup> release [patch update]: 2019-12-10_)
 ### Bug Fixes & Miscellaneous Edits
 * Speed-ups to E-step, especially for models with a noise component.
 * Clarifications and improvements to documentation and examples.
-* Fixed bug related to the colours used in the vignette plots.
 
 ## MEDseq v1.0.0 - (_1<sup>st</sup> release: 2019-08-24_)